`GitConnector` and `GitProperties` refactoring
