+++ src/main/java/ch/usi/si/seart/config/MainConfig.java
@@ -1,12 +1,9 @@
 package ch.usi.si.seart.config;
 
 import ch.usi.si.seart.bean.init.LanguageInitializationBean;
-import ch.usi.si.seart.bean.init.TemporaryDirectoryCleanerBean;
 import ch.usi.si.seart.config.properties.CrawlerProperties;
-import ch.usi.si.seart.config.properties.GitProperties;
 import ch.usi.si.seart.config.properties.StatisticsProperties;
 import ch.usi.si.seart.util.Ranges;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnExpression;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.web.servlet.FilterRegistrationBean;
@@ -16,7 +13,6 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
 import org.springframework.web.filter.ForwardedHeaderFilter;
 
-import java.nio.file.Path;
 import java.time.Instant;
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
@@ -26,16 +22,6 @@ import java.util.Date;
 
 @Configuration
 public class MainConfig {
-
-    @Bean
-    Path tmpDir(@Value("${java.io.tmpdir}") String value) {
-        return Path.of(value);
-    }
-
-    @Bean
-    public TemporaryDirectoryCleanerBean temporaryDirectoryCleanerBean(GitProperties properties, Path tmpDir) {
-        return new TemporaryDirectoryCleanerBean(tmpDir, properties.getFolderPrefix());
-    }
 
     @Bean
     @ConditionalOnProperty(value = "ghs.crawler.enabled", havingValue = "true")
