+++ src/main/java/ch/usi/si/seart/config/properties/GitProperties.java
@@ -12,7 +12,7 @@ import javax.validation.constraints.NotNull;
 import java.time.Duration;
 
 @Getter
-@ConfigurationProperties(prefix = "ghs.analysis.git", ignoreUnknownFields = false)
+@ConfigurationProperties(prefix = "ghs.git", ignoreUnknownFields = false)
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 @AllArgsConstructor(onConstructor_ = @ConstructorBinding)
 public class GitProperties {
