+++ src/main/java/ch/usi/si/seart/git/GitConnector.java
@@ -1,8 +1,6 @@
 package ch.usi.si.seart.git;
 
-import ch.usi.si.seart.config.properties.GitProperties;
 import ch.usi.si.seart.exception.TerminalExecutionException;
-import ch.usi.si.seart.exception.git.CloneException;
 import ch.usi.si.seart.exception.git.GitException;
 import ch.usi.si.seart.exception.git.RemoteReferenceDisplayException;
 import ch.usi.si.seart.io.ExternalProcess;
@@ -11,13 +9,17 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.experimental.FieldDefaults;
 import lombok.extern.slf4j.Slf4j;
+import org.apache.commons.lang3.SystemUtils;
+import org.springframework.beans.factory.InitializingBean;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.convert.ConversionService;
 
 import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import java.nio.file.Path;
+import java.time.Duration;
 import java.util.concurrent.TimeoutException;
 
 /**
@@ -27,9 +29,13 @@ import java.util.concurrent.TimeoutException;
 @Connector(command = "git")
 @AllArgsConstructor(onConstructor_ = @Autowired)
 @FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
-public class GitConnector {
+public class GitConnector implements InitializingBean {
 
-    GitProperties gitProperties;
+    @Value("${ghs.git.folder-prefix}")
+    String folderPrefix;
+
+    @Value("${ghs.git.clone-timeout-duration}")
+    Duration cloneTimeout;
 
     ConversionService conversionService;
 
@@ -44,12 +50,11 @@ public class GitConnector {
     @SuppressWarnings("ConstantConditions")
     public LocalRepositoryClone clone(URL url) throws GitException {
         try {
-            Path directory = Files.createTempDirectory(gitProperties.getFolderPrefix());
+            log.trace("Cloning:   {}", url);
+            Path directory = Files.createTempDirectory(folderPrefix);
             String[] command = {"git", "clone", "--quiet", "--depth", "1", url.toString(), directory.toString()};
             ExternalProcess process = new ExternalProcess(directory, command);
-            log.trace("Cloning:   {}", url);
-            long timeout = gitProperties.getCloneTimeoutDuration().toMillis();
-            ExternalProcess.Result result = process.execute(timeout);
+            ExternalProcess.Result result = process.execute(cloneTimeout.toMillis());
             result.ifFailedThrow(() -> {
                 GitException exception = conversionService.convert(result.getStdErr(), GitException.class);
                 return (GitException) exception.fillInStackTrace();
@@ -57,9 +62,11 @@ public class GitConnector {
             return new LocalRepositoryClone(directory);
         } catch (InterruptedException ex) {
             Thread.currentThread().interrupt();
-            throw new CloneException("Timed out for: " + url, ex);
-        } catch (IOException | TerminalExecutionException | TimeoutException ex) {
-            throw new CloneException("Failed for: " + url, ex);
+            throw new RemoteReferenceDisplayException("Cancelled for: " + url, ex);
+        } catch (TimeoutException ex) {
+            throw new RemoteReferenceDisplayException("Timed out for: " + url, ex);
+        } catch (IOException | TerminalExecutionException ex) {
+            throw new RemoteReferenceDisplayException("Failed for: " + url, ex);
         }
     }
 
@@ -79,9 +86,22 @@ public class GitConnector {
             return result.succeeded();
         } catch (InterruptedException ex) {
             Thread.currentThread().interrupt();
+            throw new RemoteReferenceDisplayException("Cancelled for: " + url, ex);
+        } catch (TimeoutException ex) {
             throw new RemoteReferenceDisplayException("Timed out for: " + url, ex);
-        } catch (TerminalExecutionException | TimeoutException ex) {
+        } catch (TerminalExecutionException ex) {
             throw new RemoteReferenceDisplayException("Failed for: " + url, ex);
+        }
+    }
+
+    @Override
+    public void afterPropertiesSet() throws Exception {
+        try {
+            Path workdir = SystemUtils.getJavaIoTmpDir().toPath();
+            new ExternalProcess(workdir, "rm", "-rf", folderPrefix + "*").execute().ifFailedThrow();
+            log.info("Successfully cleaned up repository folder");
+        } catch (TerminalExecutionException ex) {
+            log.warn("Failed to clean up cloned repositories from previous runs", ex);
         }
     }
 }
