+++ src/main/resources/application.properties
@@ -81,6 +81,10 @@ spring.flyway.password=${spring.datasource.password}
 # GSON Configuration
 spring.gson.serialize-nulls=true
 
+# Git Configuration
+ghs.git.folder-prefix=ghs-clone-
+ghs.git.clone-timeout-duration=5m
+
 # GitHub Configuration
 ghs.github.tokens=
 ghs.github.api-version=2022-11-28
@@ -115,8 +119,6 @@ ghs.crawler.delay-between-runs=PT6H
 ghs.analysis.enabled=true
 ghs.analysis.delay-between-runs=PT6H
 ghs.analysis.max-pool-threads=3
-ghs.analysis.git.folder-prefix=ghs-clone-
-ghs.analysis.git.clone-timeout-duration=5m
 ghs.analysis.cloc.analysis-timeout-duration=5m
 
 # Cleanup Configuration
